+++ CHANGES.md
@@ -1,3 +1,8 @@
+# 1.0.2
+
+* Treat all requests as `params` requests unless the operation has a representer mixed in. If you don't want that, you can override using `is_document: false`. This appears to be the smoothest solution for all.
+* In `Controller#form`, the options argument is now passed into `form.prepopulate!(options)`. This allows to use arbitrary options and the `options[:params]` for prepopulation.
+
 # 1.0.1
 
 * Treat `:js` requests as non-document, too.